requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.3
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-simplejwt==3.2.3
icalendar==4.0.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
PyJWT==1.6.1
python-dateutil==2.7.0
pytz==2018.3
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
